refactor: Update new architecture
