+++ .eslintignore
@@ -1,2 +1,2 @@
 **/node_modules/**
-/templates
+/src/templates
