+++ package.json
@@ -7,11 +7,11 @@
   "author": "Algolia <support@algolia.com>",
   "main": "index.js",
   "bin": {
-    "create-instantsearch-app": "packages/cli/index.js"
+    "create-instantsearch-app": "src/cli/index.js"
   },
   "scripts": {
-    "start": "node packages/cli/index.js",
-    "test": "jest ./packages",
+    "start": "node src/cli/index.js",
+    "test": "jest ./src",
     "test:e2e:templates": "jest './scripts/e2e-templates.test.js'",
     "test:e2e": "jest './scripts/e2e-*'",
     "lint": "eslint .",
@@ -23,8 +23,7 @@
   },
   "files": [
     "index.js",
-    "packages",
-    "templates"
+    "src"
   ],
   "engines": {
     "node": ">= 8"
@@ -62,7 +61,7 @@
   "jest": {
     "testPathIgnorePatterns": [
       "<rootDir>/node_modules/",
-      "<rootDir>/templates/"
+      "<rootDir>/src/templates/"
     ]
   }
 }
