+++ scripts/release-templates.js
@@ -15,7 +15,7 @@ const path = require('path');
 const { execSync } = require('child_process');
 const chalk = require('chalk');
 const latestSemver = require('latest-semver');
-const { fetchLibraryVersions } = require('../packages/shared/utils');
+const { fetchLibraryVersions } = require('../src/utils');
 
 const createInstantSearchApp = require('../');
 
@@ -44,7 +44,7 @@ async function build() {
     { stdio: 'ignore' }
   );
 
-  const templatesFolder = path.join(__dirname, '../templates');
+  const templatesFolder = path.join(__dirname, '../src/templates');
   const templates = fs
     .readdirSync(templatesFolder)
     .map(name => path.join(templatesFolder, name))
