+++ src/templates/React InstantSearch Native/.babelrc.template
