+++ src/templates/React InstantSearch Native/.editorconfig
