+++ src/templates/React InstantSearch Native/.eslintrc.js
