+++ src/templates/React InstantSearch Native/.gitignore.template
