+++ src/templates/React InstantSearch Native/.prettierrc
