+++ src/templates/React InstantSearch Native/.template.js
