+++ src/templates/React InstantSearch Native/.watchmanconfig
