+++ src/templates/React InstantSearch Native/App.js.hbs
