+++ src/templates/React InstantSearch Native/README.md
