+++ src/templates/React InstantSearch Native/app.json
