+++ src/templates/React InstantSearch Native/package.json
