+++ src/templates/React InstantSearch Native/src/Highlight.js
