+++ src/templates/React InstantSearch Native/src/InfiniteHits.js.hbs
