+++ src/templates/React InstantSearch Native/src/SearchBox.js.hbs
